+++ src/write_outputs/long_duration_storage/write_opwrap_lds_dstor.jl
@@ -11,7 +11,16 @@ function write_opwrap_lds_dstor(path::AbstractString, inputs::Dict, setup::Dict,
 		if i in inputs["STOR_LONG_DURATION"]
 			dsoc[i,:] = value.(EP[:vdSOC])[i,:]
 		end
+		if !isempty(inputs["VRE_STOR"])
+			if i in inputs["VS_LDS"]
+				dsoc[i,:] = value.(EP[:vdSOC_VRE_STOR])[i,:]
+			end
+		end
 	end
+	if setup["ParameterScale"] == 1
+	    dsoc *= ModelScalingFactor
+	end
+	
 	dfdStorage = hcat(dfdStorage, DataFrame(dsoc, :auto))
 	auxNew_Names=[Symbol("Resource");Symbol("Zone");[Symbol("w$t") for t in 1:W]]
 	rename!(dfdStorage,auxNew_Names)