+++ src/write_outputs/write_costs.jl
@@ -9,11 +9,27 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 	SEG = inputs["SEG"]  # Number of lines
 	Z = inputs["Z"]     # Number of zones
 	T = inputs["T"]     # Number of time steps (hours)
+	VRE_STOR = inputs["VRE_STOR"]
+	
+	cost_list = ["cTotal", "cFix", "cVar", "cNSE", "cStart", "cUnmetRsv", "cNetworkExp", "cUnmetPolicyPenalty"]
+	if !isempty(VRE_STOR)
+		push!(cost_list, "cGridConnection")
+	end
+	dfCost = DataFrame(Costs = cost_list)
 
-	dfCost = DataFrame(Costs = ["cTotal", "cFix", "cVar", "cNSE", "cStart", "cUnmetRsv", "cNetworkExp", "cUnmetPolicyPenalty"])
-	cVar = value(EP[:eTotalCVarOut])+ (!isempty(inputs["STOR_ALL"]) ? value(EP[:eTotalCVarIn]) : 0.0) + (!isempty(inputs["FLEX"]) ? value(EP[:eTotalCVarFlexIn]) : 0.0)
+	cVar = value(EP[:eTotalCVarOut]) + (!isempty(inputs["STOR_ALL"]) ? value(EP[:eTotalCVarIn]) : 0.0) + (!isempty(inputs["FLEX"]) ? value(EP[:eTotalCVarFlexIn]) : 0.0) 
 	cFix = value(EP[:eTotalCFix]) + (!isempty(inputs["STOR_ALL"]) ? value(EP[:eTotalCFixEnergy]) : 0.0) + (!isempty(inputs["STOR_ASYMMETRIC"]) ? value(EP[:eTotalCFixCharge]) : 0.0)
-	dfCost[!,Symbol("Total")] = [value(EP[:eObj]), cFix, cVar, value(EP[:eTotalCNSE]), 0.0, 0.0, 0.0, 0.0] 
+	if !isempty(VRE_STOR) 
+		cFix += ((!isempty(inputs["VS_DC"]) ? value(EP[:eTotalCFixDC]) : 0.0) + (!isempty(inputs["VS_SOLAR"]) ? value(EP[:eTotalCFixSolar]) : 0.0) + (!isempty(inputs["VS_WIND"]) ? value(EP[:eTotalCFixWind]) : 0.0))
+		cVar += ((!isempty(inputs["VS_SOLAR"]) ? value(EP[:eTotalCVarOutSolar]) : 0.0) + (!isempty(inputs["VS_WIND"]) ? value(EP[:eTotalCVarOutWind]) : 0.0))
+		if !isempty(inputs["VS_STOR"])
+			cFix += ((!isempty(inputs["VS_STOR"]) ? value(EP[:eTotalCFixStor]) : 0.0) + (!isempty(inputs["VS_ASYM_DC_CHARGE"]) ? value(EP[:eTotalCFixCharge_DC]) : 0.0) + (!isempty(inputs["VS_ASYM_DC_DISCHARGE"]) ? value(EP[:eTotalCFixDischarge_DC]) : 0.0) + (!isempty(inputs["VS_ASYM_AC_CHARGE"]) ? value(EP[:eTotalCFixCharge_AC]) : 0.0) + (!isempty(inputs["VS_ASYM_AC_DISCHARGE"]) ? value(EP[:eTotalCFixDischarge_AC]) : 0.0)) 
+			cVar += (!isempty(inputs["VS_STOR"]) ? value(EP[:eTotalCVarStor]) : 0.0)
+		end
+		dfCost[!,Symbol("Total")] = [objective_value(EP), cFix, cVar, value(EP[:eTotalCNSE]), 0.0, 0.0, 0.0, 0.0, 0.0]
+	else
+		dfCost[!,Symbol("Total")] = [objective_value(EP), cFix, cVar, value(EP[:eTotalCNSE]), 0.0, 0.0, 0.0, 0.0]
+	end
 
 	if setup["ParameterScale"] == 1
 		dfCost.Total *= ModelScalingFactor^2
@@ -46,6 +62,10 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 	if haskey(inputs, "MinCapPriceCap")
 		dfCost[8,2] += value(EP[:eTotalCMinCapSlack])
 	end	
+	
+	if !isempty(VRE_STOR)
+		dfCost[!,2][9] = value(EP[:eTotalCGrid]) * (setup["ParameterScale"] == 1 ? ModelScalingFactor^2 : 1)
+	end
 
 	if setup["ParameterScale"] == 1
 		dfCost[5,2] *= ModelScalingFactor^2
@@ -92,6 +112,72 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 			tempCVar += eCVarFlex_in
 			tempCTotal += eCVarFlex_in
 		end
+		if !isempty(VRE_STOR)
+			dfVRE_STOR = inputs["dfVRE_STOR"]
+			Y_ZONE_VRE_STOR = dfVRE_STOR[dfVRE_STOR[!,:Zone].==z,:R_ID]
+
+			# Fixed Costs
+			eCFix_VRE_STOR = 0.0
+			SOLAR_ZONE_VRE_STOR = intersect(Y_ZONE_VRE_STOR, inputs["VS_SOLAR"])
+			if !isempty(SOLAR_ZONE_VRE_STOR)
+				eCFix_VRE_STOR += sum(value.(EP[:eCFixSolar][SOLAR_ZONE_VRE_STOR]))
+			end
+			WIND_ZONE_VRE_STOR = intersect(Y_ZONE_VRE_STOR, inputs["VS_WIND"])
+			if !isempty(WIND_ZONE_VRE_STOR)
+				eCFix_VRE_STOR += sum(value.(EP[:eCFixWind][WIND_ZONE_VRE_STOR]))
+			end
+			DC_ZONE_VRE_STOR = intersect(Y_ZONE_VRE_STOR, inputs["VS_DC"])
+			if !isempty(DC_ZONE_VRE_STOR)
+				eCFix_VRE_STOR += sum(value.(EP[:eCFixDC][DC_ZONE_VRE_STOR]))
+			end
+			STOR_ALL_ZONE_VRE_STOR = intersect(inputs["VS_STOR"], Y_ZONE_VRE_STOR)
+			if !isempty(STOR_ALL_ZONE_VRE_STOR)
+				eCFix_VRE_STOR += sum(value.(EP[:eCFixEnergy_VS][STOR_ALL_ZONE_VRE_STOR]))
+				DC_CHARGE_ALL_ZONE_VRE_STOR = intersect(inputs["VS_ASYM_DC_CHARGE"], Y_ZONE_VRE_STOR)
+				if !isempty(DC_CHARGE_ALL_ZONE_VRE_STOR)
+					eCFix_VRE_STOR += sum(value.(EP[:eCFixCharge_DC][DC_CHARGE_ALL_ZONE_VRE_STOR]))
+				end
+				DC_DISCHARGE_ALL_ZONE_VRE_STOR = intersect(inputs["VS_ASYM_DC_DISCHARGE"], Y_ZONE_VRE_STOR)
+				if !isempty(DC_DISCHARGE_ALL_ZONE_VRE_STOR)
+					eCFix_VRE_STOR += sum(value.(EP[:eCFixDischarge_DC][DC_DISCHARGE_ALL_ZONE_VRE_STOR]))
+				end
+				AC_DISCHARGE_ALL_ZONE_VRE_STOR = intersect(inputs["VS_ASYM_AC_DISCHARGE"], Y_ZONE_VRE_STOR)
+				if !isempty(AC_DISCHARGE_ALL_ZONE_VRE_STOR)
+					eCFix_VRE_STOR += sum(value.(EP[:eCFixDischarge_AC][AC_DISCHARGE_ALL_ZONE_VRE_STOR]))
+				end
+				AC_CHARGE_ALL_ZONE_VRE_STOR = intersect(inputs["VS_ASYM_AC_CHARGE"], Y_ZONE_VRE_STOR)
+				if !isempty(AC_CHARGE_ALL_ZONE_VRE_STOR)
+					eCFix_VRE_STOR += sum(value.(EP[:eCFixCharge_AC][AC_CHARGE_ALL_ZONE_VRE_STOR]))
+				end
+			end
+			tempCFix += eCFix_VRE_STOR
+
+			# Variable Costs
+			eCVar_VRE_STOR = 0.0
+			if !isempty(SOLAR_ZONE_VRE_STOR)
+				eCVar_VRE_STOR += sum(value.(EP[:eCVarOutSolar][SOLAR_ZONE_VRE_STOR,:]))
+			end
+			if !isempty(WIND_ZONE_VRE_STOR)
+				eCVar_VRE_STOR += sum(value.(EP[:eCVarOutWind][WIND_ZONE_VRE_STOR, :]))
+			end
+			if !isempty(STOR_ALL_ZONE_VRE_STOR)
+				vom_map = Dict(
+					DC_CHARGE_ALL_ZONE_VRE_STOR => :eCVar_Charge_DC,
+					DC_DISCHARGE_ALL_ZONE_VRE_STOR => :eCVar_Discharge_DC,
+					AC_DISCHARGE_ALL_ZONE_VRE_STOR => :eCVar_Discharge_AC,
+					AC_CHARGE_ALL_ZONE_VRE_STOR => :eCVar_Charge_AC
+				)
+				for (set, symbol) in vom_map
+					if !isempty(set)
+						eCVar_VRE_STOR += sum(value.(EP[symbol][set, :]))
+					end
+				end
+			end
+			tempCVar += eCVar_VRE_STOR
+
+			# Total Added Costs
+			tempCTotal += (eCFix_VRE_STOR + eCVar_VRE_STOR)
+		end
 
 		if setup["UCommit"] >= 1 && !isempty(COMMIT_ZONE)
 			eCStart = sum(value.(EP[:eCStart][COMMIT_ZONE,:]))
@@ -109,7 +195,11 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 			tempCNSE *= ModelScalingFactor^2
 			tempCStart *= ModelScalingFactor^2
 		end
-		dfCost[!,Symbol("Zone$z")] = [tempCTotal, tempCFix, tempCVar, tempCNSE, tempCStart, "-", "-", "-"]
+		temp_cost_list = [tempCTotal, tempCFix, tempCVar, tempCNSE, tempCStart, "-", "-", "-"]
+		if !isempty(VRE_STOR)
+			push!(temp_cost_list, "-")
+		end
+		dfCost[!,Symbol("Zone$z")] = temp_cost_list
 	end
 	CSV.write(joinpath(path, "costs.csv"), dfCost)
 end